bottom navbar works properly
