+++ lib/presentation/pages/tips.dart
@@ -1,0 +1,19 @@
+import 'package:flutter/material.dart';
+import 'package:go_router/go_router.dart';
+import 'package:google_nav_bar/google_nav_bar.dart';
+import 'package:wodobro/presentation/home_page.dart';
+
+class TipsPage extends StatefulWidget {
+  const TipsPage({Key? key}) : super(key: key);
+
+  @override
+  State<TipsPage> createState() => _TipsPageState();
+}
+
+class _TipsPageState extends State<TipsPage> {
+  //const TipsPage({Key? key}) : super(key: key);
+  @override
+  Widget build(BuildContext context) {
+    return Center(child: Text('Tips Page'));
+  }
+}
