--- lib/presentation/tips_page.dart
@@ -1,42 +1,0 @@
-import 'package:flutter/material.dart';
-import 'package:go_router/go_router.dart';
-import 'package:wodobro/presentation/home_page.dart';
-
-class TipsPage extends StatelessWidget {
-  const TipsPage({Key? key}) : super(key: key);
-
-  final selectedPageIndex = 1;
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-      body: Center(child: Text('Tips Page')),
-      bottomNavigationBar: NavigationBar(
-        selectedIndex: selectedPageIndex,
-        onDestinationSelected: (int index) {
-          if (index == 0) {
-            Navigator.pushReplacement(
-              context,
-              PageRouteBuilder(
-                pageBuilder: (context, animation1, animation2) => HomePage(),
-                transitionDuration: Duration.zero,
-                reverseTransitionDuration: Duration.zero,
-              ),
-            );
-          }
-        },
-        destinations: const <NavigationDestination>[
-          NavigationDestination(
-            selectedIcon: Icon(Icons.home),
-            icon: Icon(Icons.home_outlined),
-            label: 'Home',
-          ),
-          NavigationDestination(
-            selectedIcon: Icon(Icons.lightbulb),
-            icon: Icon(Icons.lightbulb_outlined),
-            label: 'Tips',
-          ),
-        ],
-      ),
-    );
-  }
-}
