+++ pubspec.yaml
@@ -43,6 +43,7 @@ dependencies:
   google_fonts: ^4.0.4
   animate_icons: 2.0.0
   intl: ^0.18.1
+  google_nav_bar: ^5.0.6
   circular_chart_flutter: ^0.0.2
   json_annotation: ^4.8.0
   firebase_core: ^2.13.0
